+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project starting with the 0.6.0 release will be docu
 - Password no longer shows when logging in with some Windows terminal clients (#574)
 - No more errors when running Terminus in Windows from directories with spaces in the path. (#575)
 - `site create-env` no longer fails to clone from an environment. (#602)
+- `site backups list` filtering by element fixed. (#602)
 
 ### Changed
 - Logged errors now exit with -1. (#576)